Refactor search logic in the "find" page into its own component
- Separate the logic that lets a user search for locations
  into its own component, as the MapPage is getting pretty big.
